requirements.txt
black==23.3.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
h3==3.7.6
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
psycopg2==2.9.6
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
SQLAlchemy==2.0.15
timezonefinder==6.2.0
tomli==2.0.1
typing_extensions==4.6.3
urllib3==2.0.2
Werkzeug==2.3.4
